feat: add univ select page to verify steps
