+++ src/pages/common/verifyForMatchingResultStep/Step.tsx
@@ -1,11 +1,12 @@
 import PageLayout from '~/components/layout/page/PageLayout';
 import { useFunnel } from '~/hooks/useFunnel';
 import FirstPage from './FirstPage';
+import SecondPage from './SecondPage';
 
-const PAGE_NUMBER = [1];
+const PAGE_NUMBER = [1, 2];
 
 const CommonVerifyForMatchingResultStep = () => {
-  const { Funnel, PageHandler } = useFunnel({
+  const { Funnel, PageHandler, currentPage } = useFunnel({
     pageNumberList: PAGE_NUMBER,
     prevStep: { path: '/' },
     nextStep: { path: '/' }, // TODO: 시대팅 안내 사항 보러가기 페이지로 수정
@@ -19,11 +20,14 @@ const CommonVerifyForMatchingResultStep = () => {
           <Funnel.Page pageNumber={1}>
             <FirstPage />
           </Funnel.Page>
+          <Funnel.Page pageNumber={2}>
+            <SecondPage />
+          </Funnel.Page>
         </Funnel>
       </PageLayout.SingleCardBody>
       <PageLayout.Footer
-        currentPage={1}
-        totalPage={1}
+        currentPage={currentPage}
+        totalPage={2}
         onPrev={PageHandler.onPrev}
         onNext={PageHandler.onNext}
       />
